+++ Data.txt
@@ -1,3 +1,3 @@
-[D][ ] event4 (by: Oct 15 2019)
-[E][ ] fair (from: Oct 15 2019 to Oct 16 2019)
-[E][ ] student fair (from: Oct 15 2019 to Oct 17 2019)
+[T][ ] list
+[D][ ] list (by: Oct 15 2019)
+[E][ ] list (from: Oct 15 2019 to Oct 16 2019)
