+++ src/main/java/Task.java
@@ -22,8 +22,4 @@ public class Task {
     public String toString() {
         return "[" + (isDone ? "X" : " ") + "] " + description;
     }
-
-    public String getType() {
-        return "";
-    }
 }
