+++ src/main/java/Ui.java
@@ -1,2 +1,19 @@
-package PACKAGE_NAME;public class Ui {
+public class Ui {
+
+    public static final String line = "____________________________________________________________";
+
+    public Ui() {
+    }
+
+    public static void printHello() {
+        System.out.println(line);
+        System.out.println("Hello! I am TaskMaster!");
+        System.out.println("What can I do for you today?");
+        System.out.println(line);
+    }
+    public static void printGoodbye() {
+        System.out.println(line);
+        System.out.println("Bye! Hope to see you again!");
+        System.out.println(line);
+    }
 }
